moved the classes Config and Statement into seperate files
